+++ README.rst
@@ -51,7 +51,7 @@ Write your first tests file to `test_myinfra.py`:
 
 And run it::
 
-    $ testinfra -v test_myinfra.py
+    $ py.test -v test_myinfra.py
 
 
     ====================== test session starts ======================
